pyproject.toml
[tool.poetry]
name = "haystack-udemy"
version = "0.1.0"
description = ""
authors = ["raulest50 <raulest50@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
haystack-ai = "^2.6.1"
datasets = "^3.1.0"
sentence-transformers = "^3.2.1"
wikipedia = "^1.4.0"
qdrant-haystack = "^7.0.0"
chroma-haystack = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

